+++ packages/client/src/components/Game/entities/border.ts
