+++ packages/client/src/components/Game/entities/ghost.ts
