+++ packages/client/src/components/Game/entities/pac-man.ts
