+++ packages/client/src/components/Game/entities/utils.ts
